+++ KlayGE/Plugins/Src/Render/OpenGL/OGLTexture3D.cpp
@@ -66,9 +66,9 @@ namespace KlayGE
 			{
 				++ num_mip_maps_;
 
-				w = std::max(static_cast<uint32_t>(1), w / 2);
-				h = std::max(static_cast<uint32_t>(1), h / 2);
-				d = std::max(static_cast<uint32_t>(1), d / 2);
+				w = std::max(1U, w / 2);
+				h = std::max(1U, h / 2);
+				d = std::max(1U, d / 2);
 			}
 		}
 		else
@@ -127,25 +127,25 @@ namespace KlayGE
 					(NULL == init_data) ? NULL : init_data[level].data);
 			}
 
-			width = std::max(static_cast<uint32_t>(1), width / 2);
-			height = std::max(static_cast<uint32_t>(1), height / 2);
-			depth = std::max(static_cast<uint32_t>(1), depth / 2);
+			width = std::max(1U, width / 2);
+			height = std::max(1U, height / 2);
+			depth = std::max(1U, depth / 2);
 		}
 
 		this->UpdateParams();
 	}
 
-	uint32_t OGLTexture3D::Width(int level) const
+	uint32_t OGLTexture3D::Width(uint32_t level) const
 	{
 		return widthes_[level];
 	}
 
-	uint32_t OGLTexture3D::Height(int level) const
+	uint32_t OGLTexture3D::Height(uint32_t level) const
 	{
 		return heights_[level];
 	}
 
-	uint32_t OGLTexture3D::Depth(int level) const
+	uint32_t OGLTexture3D::Depth(uint32_t level) const
 	{
 		return depthes_[level];
 	}
